requirements.txt
# ============================================================
# Bangladeshi Education AI – Python Backend Dependencies
# ============================================================

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Settings management
pydantic==2.9.2
pydantic-settings==2.5.2

# Async MongoDB driver
motor==3.6.0
pymongo==4.10.1

# Redis cache
redis==5.1.1

# ChromaDB vector store (RAG)
chromadb==0.5.15

# AI SDKs
google-generativeai==0.8.3
openai==1.50.2          # used for DeepSeek and Grok (OpenAI-compatible)

# Image processing (Data Saver mode)
Pillow==11.1.0

# JWT authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0          # retry logic
python-multipart==0.0.12 # form data support
